Change the way functions are randomly seeded
Created a function decorator which essentially just adds the
keyword argument `seed` to the function it is decorating. This
integer value is supplied to numpy: `np.random.seed(seed)`, and
the function being decorated has become reproducible :smile:.

To decorate a function, just add @set_random_seed before it, and it
will automagically take `seed` as a keyword argument.