Added new pianoroll class
The point of this class is to define what a pianoroll is. I have
defined a pianoroll as an arbitrary matrix representation of a
composition. It will be able to store different attributes such
as note on events, note off events, and pitch sounding events. It
can also handle having multiple tracks.

The main this is that, after its construction with a quantized
note dataframe, it doesn't hold onto this dataframe, but rather
is just a storage container for the matrices that represent this
dataframe. At minimum it stores onsets and sounding events, so
there is no loss of information.

Most of the time we might only be interested in using a subset of
the possible channels stored. The __call__ method allows for
arbitrary selection of tracks and channels.