Allow bypass of overlap check
Firstly, this overlap check may not always be strictly necessary
(I sort of think it is - it shouldn't be possible to start playing
a note at a given pitch before a previous note at the same pitch
has finished IMHO!...maybe in the case of weird synths this could
be relaxed!). Secondly, the check is the main contributor to dog
slow time. Removing it speeds everything up by about 100x. Should
investigate the holdup.