Made methods importable and added args
Made it such that the main call can be executed in other scripts.
I needed this for collating the evaluation metrics in another script.
In conjuction with the fact that trainiers now add eval info to
the log information, we can now programatically evaluate and store
evaluations from a list of models

Also added a parameter --splits allowing you to perform the evaluation
on training and validation splits too.